PLANET-4811 Create SubmenuItems component and useSubmenuItemsLoad hook
