+++ assets/src/blocks/Submenu/submenuFunctions.js
@@ -75,45 +75,9 @@ const addTargetLinks = item => {
   }
 };
 
-export const loadMenuItems = (levels, isEditing) => {
-  // Get all heading tags that we need to query
-  const headings = levels.map(level => `h${level.heading}`);
-  const tags = getHeadings(headings, isEditing ? 'editor-styles-wrapper' : 'page-template');
-  if (!tags) {
-    return [];
-  }
-  return tags.reduce((menuItems, tag, index) => {
-    const headingNumber = getHeadingNumber(tag);
-    let previousHeadingNumber = 0;
-    if (index > 0) {
-      previousHeadingNumber = getHeadingNumber(tags[index - 1]);
-    }
-    // Get the properties that we need to create the new menu item
-    const correspondingLevel = levels.find(level => level.heading === headingNumber);
-    const id = tag.id || tag.textContent.toLowerCase().replace(/ /g, '-').replace(/[^\w-]+/g, ''); // equivalent of WP sanitize_title function
-    const menuItem = {
-      text: tag.textContent,
-      id,
-      style: correspondingLevel.style,
-      link: correspondingLevel.link,
-      type: `h${correspondingLevel.heading}`,
-      children: []
-    };
-    if (previousHeadingNumber && headingNumber > previousHeadingNumber) {
-      // In this case we need to add this menuItem to the children of the previous one
-      menuItems[menuItems.length - 1].children.push(menuItem);
-    } else {
-      menuItems.push(menuItem);
-    }
-    return menuItems;
-  }, []);
-};
-
-const getHeadings = (headings, className = 'page-template') => {
+export const getHeadings = (headings, className = 'page-template') => {
   // We need to make sure it's a div element,
   // since for 'page-template' className we have it on the body too
   const page = document.querySelector(`div.${className}`);
   return page ? [...page.querySelectorAll(headings)] : null;
 };
-
-const getHeadingNumber = tag => Number(tag.tagName.replace('H', ''));
